Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -6,7 +6,7 @@ int main(){
     scanf("%d", &b);
     int c;
     scanf("%d", &c);
-    int d = (a+b+c)/3
-    printf("You entered: %d",d);
+    float d = (a+b+c)/3
+    printf("You entered: %.2f",d);
     return 0;
 }
